+++ ydb/core/kqp/executer_actor/kqp_executer.h
@@ -1,12 +1,13 @@
 #pragma once
 
+#include "shards_resolver/kqp_shards_resolver_events.h"
+#include "tasks_graph/kqp_partition_helper.h"
+
 #include <library/cpp/lwtrace/shuttle.h>
 #include <ydb/core/kqp/common/kqp_batch_operations.h>
 #include <ydb/core/kqp/common/kqp_tx.h>
 #include <ydb/core/kqp/common/kqp_event_ids.h>
 #include <ydb/core/kqp/common/kqp_user_request_context.h>
-#include <ydb/core/kqp/executer_actor/kqp_partition_helper.h>
-#include <ydb/core/kqp/executer_actor/shards_resolver/kqp_shards_resolver_events.h>
 #include <ydb/core/kqp/query_data/kqp_query_data.h>
 #include <ydb/core/kqp/gateway/kqp_gateway.h>
 #include <ydb/core/kqp/counters/kqp_counters.h>
